requirements.txt
absl-py==0.13.0
aiocontextvars==0.2.2
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
async-timeout==3.0.1
attrs==21.2.0
binary-search==0.3.0
blessings==1.7
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.1
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
filelock==3.0.12
google-api-core==1.31.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.38.1
gym==0.18.3
h5py==3.1.0
hiredis==2.0.0
idna==3.2
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.6.1
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
llvmlite==0.36.0
loguru==0.5.3
Markdown==3.3.4
matplotlib==3.3.4
msgpack==1.0.2
multidict==5.1.0
networkx==2.5.1
numba==0.53.1
numpy==1.17.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.5.2.54
packaging==21.0
pandas==1.1.5
Pillow==8.2.0
prometheus-client==0.11.0
protobuf==3.17.3
psutil==5.8.0
py-spy==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyglet==1.5.15
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
ray==1.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
simpy==4.0.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
tianshou==0.4.2
torch==1.9.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
utm==0.7.0
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
